refactor: move components directory to project root
File Structure:

- Relocate `app/components/` to root-level `components/`
- Update imports from `@/app/components/` to `@/components/`

Aligns with shadcn/ui convention and keeps `app/` focused on routing. Shorter
import paths and better ecosystem compatibility.
