+++ app/layout.tsx
@@ -2,7 +2,7 @@ import { Analytics } from "@vercel/analytics/next";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import type { Metadata } from "next";
 import "./globals.css";
-import { ThemeProvider } from "@/app/components/theme-provider";
+import { ThemeProvider } from "@/components/theme-provider";
 import { geistSans, jetBrainsMono, spaceGrotesk } from "./fonts";
 
 export const metadata: Metadata = {
